fix 合并字图

+++ controller/data/api.py
@@ -140,10 +140,16 @@ class VariantMergeApi(BaseHandler):
         try:
             rules = [(v.not_empty, 'img_names', 'main')]
             self.validate(self.data, rules)
-            assert self.data['main'] in self.data['img_names']
+            img_names, main = self.data['img_names'], self.data['main']
+            assert main in img_names
+            names2merge = [name for name in img_names if name != main]
+            variants = list(self.db.variant.find({'img_name': {'$in': img_names}}))
+            if not names2merge or not variants:
+                return self.send_error_response(e.no_object, message='没有找到待合并的异体字字图')
+
             # 更新字数据
-            names2merge = [name for name in self.data['img_names'] if name != self.data['main']]
-            self.db.char.update_many({'txt': {'$in': names2merge}}, {'$set': {'txt': self.data['main']}})
+            maps = {t['img_name']: 'Y%s' % t['uid'] for t in variants}
+            self.db.char.update_many({'txt': {'$in': [maps[n] for n in names2merge]}}, {'$set': {'txt': maps[main]}})
             # 删除异体字图
             self.db.variant.delete_many({'img_name': {'$in': names2merge}})
             self.send_data_response()
